fix bug: argadrr returns void not int :))

+++ kernel/sysinfo.c
@@ -7,7 +7,8 @@
 
 #include "sysinfo.h"
 
-int get_sysinfo(uint64 addr)
+int
+get_sysinfo(uint64 addr)
 {
     // get the current proc
     // Return the current struct proc *, or zero if none.
@@ -23,7 +24,7 @@ int get_sysinfo(uint64 addr)
      * cur_proc->pagetable: page table of the calling process
      * (char *)&info: kernel-space ptr to the sysinfo data to be copied
     */
-    if(copyout(cur_proc->pagetable, addr, (char*)&info, sizeof(info)) < 0)
+    if(copyout(cur_proc->pagetable, (uint64)addr, (char*)&info, sizeof(info)) < 0)
     {
         return -1; //error
     }
